三角形と点の判定完了

+++ Headers/Judge.h
@@ -15,7 +15,7 @@ public:
 	Judge(){};
 	~Judge();
 
-	static bool c_b(const Object3d& capsle,const int height, const Object3d& sphere);	// Object3dの値で頑張ろうとしてるやつ
+	static bool c_b(const Object3d& capsle,const float height, const Object3d& sphere);	// Object3dの値で頑張ろうとしてるやつ
 	static bool JudgeCap_Sphe(const Geometric_Capsule& capsle, const Geometric_Sphere& sphere);	// Geomteric_[X]を使って判定しようとしたやつ
 	static void HitJudge(Player* p, Enemy* e);	// 当たり判定を取る、ここで処理を行う
